server/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic[email]
passlib
python-jose
python-jose[cryptography]
